[#11]: Refactored interfaces into their own files
