+++ Server/src/types/client_event.ts
@@ -1,0 +1,5 @@
+export interface ClientEvent {
+    funcName: string;  // Todo: rename to "function"
+    params: Array<string | number | boolean>;
+    executeCycleNumber: number;
+}
