+++ Server/src/types/command.ts
@@ -1,0 +1,6 @@
+export type Commands = Record<string, Command>;
+
+export interface Command {
+    funcName: string;
+    params: Array<string | number | boolean>;
+}
