+++ Server/src/types/steam.ts
@@ -1,20 +1,20 @@
 export interface SteamPlayerSummary {
-    steamid: string;
-    communityvisibilitystate: number;
-    profilestate: number;
-    personaname: string;
-    commentpermission: number;
-    profileurl: string;
     avatar: string;
-    avatarmedium: string;
     avatarfull: string;
     avatarhash: string;
+    avatarmedium: string;
+    commentpermission: number;
+    communityvisibilitystate: number;
     lastlogoff: number;
-    personastate: number;
-    primaryclanid: string;
-    timecreated: number;
-    personastateflags: number;
     loccountrycode: string;
+    personaname: string;
+    personastate: number;
+    personastateflags: number;
+    primaryclanid: string;
+    profilestate: number;
+    profileurl: string;
+    steamid: string;
+    timecreated: number;
 }
 
 export interface SteamPlayerSummeryResponse {
